Rename Support namespace to Utility
It is a "Utility class to provide string checks and manipulations" after all.
