+++ src/main/php/PHPMD/Utility/Strings.php
@@ -15,7 +15,7 @@
  * @link http://phpmd.org/
  */
 
-namespace PHPMD\Support;
+namespace PHPMD\Utility;
 
 use InvalidArgumentException;
